Use TypeScript; update deps; refactor exports
* **(BREAKING)** Change `index` so that it exports all public modules as
  named exports
* **(BREAKING)** Update all dependencies so that we get TypeScript
  versions
* Convert implementation code to TypeScript
* Convert tests to TypeScript, and add tests to ensure that the
  JavaScript-only checks still work for as much backward compatibility
  as possible
* Add a working TypeDoc configuration file in the event that we want to
  generate documentation
