+++ .gitignore
@@ -38,6 +38,9 @@ node_modules/
 # TypeScript cache
 *.tsbuildinfo
 
+# TypeDoc docs
+typedocs/
+
 # Optional npm cache directory
 .npm
