+++ package.json
@@ -7,12 +7,13 @@
     "url": "https://github.com/MetaMask/eth-json-rpc-infura.git"
   },
   "license": "ISC",
-  "main": "src/index.js",
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
   "files": [
-    "src/*.js"
+    "dist/"
   ],
   "scripts": {
-    "build": "echo 'nothing to build yet'",
+    "build": "tsc --project tsconfig.build.json",
     "build:clean": "rimraf dist && yarn build",
     "lint": "yarn lint:eslint && yarn lint:misc --check",
     "lint:eslint": "eslint . --cache --ext js,ts",
@@ -24,9 +25,9 @@
     "test:watch": "jest --watch"
   },
   "dependencies": {
-    "eth-json-rpc-middleware": "^6.0.0",
-    "eth-rpc-errors": "^3.0.0",
-    "json-rpc-engine": "^5.3.0",
+    "eth-json-rpc-middleware": "^8.0.1",
+    "eth-rpc-errors": "^4.0.3",
+    "json-rpc-engine": "^6.1.0",
     "node-fetch": "^2.6.7"
   },
   "devDependencies": {
@@ -35,8 +36,15 @@
     "@metamask/eslint-config": "^9.0.0",
     "@metamask/eslint-config-jest": "^9.0.0",
     "@metamask/eslint-config-nodejs": "^9.0.0",
+    "@metamask/eslint-config-typescript": "^9.0.1",
+    "@types/jest": "^26.0.13",
+    "@types/node": "^17.0.23",
+    "@types/node-fetch": "^2.6.1",
+    "@typescript-eslint/eslint-plugin": "^4.21.0",
+    "@typescript-eslint/parser": "^4.21.0",
     "eslint": "^7.23.0",
     "eslint-config-prettier": "^8.5.0",
+    "eslint-import-resolver-typescript": "^2.7.0",
     "eslint-plugin-import": "^2.25.4",
     "eslint-plugin-jest": "^24.3.4",
     "eslint-plugin-jsdoc": "^36.1.0",
@@ -45,7 +53,11 @@
     "jest": "^26.4.2",
     "prettier": "^2.6.1",
     "prettier-plugin-packagejson": "^2.2.17",
-    "rimraf": "^3.0.2"
+    "rimraf": "^3.0.2",
+    "ts-jest": "^26.3.0",
+    "ts-node": "^10.7.0",
+    "typedoc": "^0.22.13",
+    "typescript": "~4.4.0"
   },
   "engines": {
     "node": ">=12.0.0"
@@ -57,8 +69,8 @@
   "lavamoat": {
     "allowScripts": {
       "@lavamoat/preinstall-always-fail": false,
-      "eth-json-rpc-middleware>ethereumjs-util>ethereum-cryptography>keccak": true,
-      "eth-json-rpc-middleware>ethereumjs-util>ethereum-cryptography>secp256k1": true
+      "eth-json-rpc-middleware>eth-sig-util>ethereumjs-util>ethereum-cryptography>keccak": true,
+      "eth-json-rpc-middleware>eth-sig-util>ethereumjs-util>ethereum-cryptography>secp256k1": true
     }
   }
 }
