+++ src/index.ts
@@ -1,0 +1,6 @@
+export {
+  default as createInfuraMiddleware,
+  CreateInfuraMiddlewareOptions,
+} from './createInfuraMiddleware';
+export { default as fetchConfigFromReq } from './fetchConfigFromReq';
+export { default as createProvider } from './createProvider';
